Fix docs for running conan_server in a subdirectory
The previous suggested configuration works OK for search, but it fails on
upload.

Without a trailing slash in the remote, the urljoin implementation drops the
subdirectory from the URL created for http PUTs.

The suggested nginx configuration did not work for upload either. It fails to
preserve the query string, meaning the jwt token is lost. The technique here
should preserve all relevant information.