MSBuild use_env parameter: Clarify behavior
At first glance, the `use_env` parameter seems to imply the intuitive idea that no environment variables will be used in a build (but it doesn't).  I was the one who originally requested the feature, and a year later, even I thought this would be the case.  It's not Conan's fault, the parameter directly maps to the equivalent `MSBuild` flag/parameter and that makes sense.  The confusion comes from believing this flag maps to some internal Conan behavior, in which case one might expect the `MSBuild` helper to skip the application of the environment variables to the process.  But, since it doesn't, this clarification should be helpful to those trying to understand what's happening while debugging.