change all generator names/refs to lowercase (#1894)
There was one actual error on this doc which has been corrected (the class had uppercase characters but the instructions with the reference was all lowercase. Made class name lowercase to match recommended convention.  Also, updated several other items in the document to be lowercase and/or add cosmetic improvements.