Fix CLI example for creating a new package from a template (#2500)
The command line as is produces this error:
```
conan new: error: unrecognized arguments: --template=cmake_lib
```

I had to read the code to figure out what the problem was. In particular,
the command `conan new cmake_lib` produces a cryptic error form Jinja.
It would be good to have a better error output from the command as well.