Conan Tutorial Docs (#3570)
* Update build_simple_cmake_project.rst

More natural native English.

* Update different_configurations.rst

1. Changed typo 'or' to 'of'.
2. Reword section on conan install installing debug versions of libraries. Suggest that using the general meaning of installing while talking about the conan install command is confusing so hopefully it is more obvious to be explicit about the local cache.
3. Remove caption 'Windows' from platform agnostic conan command.

* Update create_your_first_package.rst

Reword description of layout method.  A sentence starting with 'Things like', makes it difficult to detect the verb and one is not sure whether it refers back to the previous sentence.

* Update add_dependencies_to_packages.rst

Remove a 'that' and replace with 'to how' for more natural English.

* Update preparing_the_build.rst

Insert missing word 'how'.
'We explain to use' becomes 'We explain how to use'.

* Update configure_options_settings.rst

'constraint' is a noun - 'to constraint' replaced with 'to constrain'.

* Update configure_options_settings.rst

Reword the  'Be aware' paragraph; it was not clear that the 'or' was mutually exclusive.
Reword the phrasing about an exception being raised though it is still not clear whether the warning mentioned at the end of the paragraph is the same as the exception or not.
Replace 'where conditions meet' with 'where conditions are met' for more idiomatic English.

* Update configure_options_settings.rst

Replace 'This steps' with 'These steps' - typo.

* Update build_packages.rst

Change text about testing. It was confusing, think is says what it was meant to say now.
Add some missing words in other sections.

* Update tutorial/creating_packages/build_packages.rst

Co-authored-by: James <memsharded@gmail.com>

* Update tutorial/creating_packages/build_packages.rst

---------

Co-authored-by: Rubén Rincón Blanco <git@rinconblanco.es>
Co-authored-by: James <memsharded@gmail.com>