Clarify docs for version-range carat operator
The docs say that variability is allowed in the "last defined digit",
which suggests that "^1.0.0" should include only versions of the form
"1.0.X", but actually at accepts "1.X.X". The actual logic doesn't have
anything to do with the "last defined" digit of any kind. Instead, it
looks for the first non-zero digit, and allows the following digit to
vary.